requirements.txt
python-telegram-bot[job-queue]
requests
pymongo
python-dotenv
fastapi
uvicorn
httpx
